ipacm: changes related to IP Passhthrough
In IP passthrough, WAN interface will be assigned a private ip
address. NAT rules related this IP need to be added to the HW
so that there will not be any exceptions. Give provision in
XML using which QCMAP indicates passthrough mode. Whenever
passthrough mode is enabled and the private ip matches with
wan ip, nat rules will be added to the HW.

Change-Id: I7e82b6f55391da20f5c5db9e00bc83508b7f8788