Translate TRY_AGAIN error for SetQuota() to success inline with HAL definition
In sequence of events where interface is down on modem side and framework is
not aware of this, FW tries to setQuota on interface which is already down.

Q6 returns failure in this situation and framework stops IPA OFFLOAD on failure.
Until tethering is switched off and on again, packets take SW path.

To avoid this, IPACM sends TRY_AGAIN to HAL and HAL forward the success
message to framework. Once upstream come up again, setQuota will be trigerred
again with proper paramaters.

Change-Id: I73e5ddbdedbc2ec006350a59d5795056dc1271b0
CRs-Fixed: 2142519